refactor most commands into their own modules
