+++ src/errors.rs
@@ -4,6 +4,4 @@ use thiserror::Error;
 pub(crate) enum UserError {
 	#[error("User aborted the operation.")]
 	Aborted,
-	#[error("Unknown subcommand. It may need to be implemented.")]
-	UnknownSubcommand,
 }
